+++ torchcam/__init__.py
@@ -1,4 +1,4 @@
-from torchcam import cams, utils
+from torchcam import methods, utils
 
 try:
     from .version import __version__  # noqa: F401
