--- torchcam/cams/__init__.py
@@ -1,3 +1,0 @@
-from .cam import *
-from .gradcam import *
-from .utils import *
