+++ torchcam/methods/__init__.py
@@ -1,0 +1,2 @@
+from .activation import *
+from .gradient import *
