+++ torchcam/methods/core.py
@@ -11,7 +11,7 @@ import torch
 import torch.nn.functional as F
 from torch import Tensor, nn
 
-from .utils import locate_candidate_layer
+from ._utils import locate_candidate_layer
 
 __all__ = ['_CAM']
